SmlAdminDataCalculator moved to spark.calculations package
