+++ src/main/scala/spark/calculations/SmlAdminDataCalculator.scala
@@ -1,6 +1,7 @@
-package uk.gov.ons.registers.methods
+package spark.calculations
 
 import org.apache.spark.sql.{DataFrame, SparkSession}
+import uk.gov.ons.registers.methods.{PayeCalculator, VatCalculator}
 
 trait SmlAdminDataCalculator extends PayeCalculator with VatCalculator{
